Fix - navibar 위치조정
