+++ frontend/app/src/layouts/MobileFrame.tsx
@@ -1,5 +1,11 @@
 import React, { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
+import { useNavigate, useLocation } from "react-router-dom";
+import { COLORS } from '../styles/colors';
+
+// NavigationBar 아이콘 import
+import mypageIcon from '../assets/icons/navigation/mypage.png';
+import calendarIcon from '../assets/icons/navigation/calendar.png';
+import reportIcon from '../assets/icons/navigation/report.png';
 
 interface MobileFrameProps {
   children: React.ReactNode;
@@ -15,11 +21,40 @@ const SCREEN_RADIUS = 38;
 
 export default function MobileFrame({ children }: MobileFrameProps) {
   const navigate = useNavigate();
+  const location = useLocation();
   
   const [time, setTime] = useState(() => {
     const now = new Date();
     return now.toLocaleTimeString("ko-KR", { hour: "2-digit", minute: "2-digit", hour12: false });
   });
+
+  // NavigationBar를 표시할 페이지들
+  const pagesWithNavigation = ['/mypage', '/calendar'];
+  const showNavigation = pagesWithNavigation.includes(location.pathname);
+
+  // NavigationBar 탭 설정
+  const navigationTabs = [
+    {
+      id: 'my',
+      label: '마이',
+      path: '/mypage',
+      icon: mypageIcon
+    },
+    {
+      id: 'calendar',
+      label: '캘린더', 
+      path: '/calendar',
+      icon: calendarIcon
+    },
+    {
+      id: 'reports',
+      label: '리포트',
+      path: '/reports', 
+      icon: reportIcon
+    }
+  ];
+
+  const isActive = (path: string) => location.pathname === path;
 
   useEffect(() => {
     const interval = setInterval(() => {
@@ -124,6 +159,63 @@ export default function MobileFrame({ children }: MobileFrameProps) {
       >
         <div style={{ height: "100%", overflowY: "auto" }}>
           {children}
+          
+          {/* NavigationBar - 특정 페이지에서만 표시 */}
+          {showNavigation && (
+            <div style={{
+              position: 'absolute',
+              bottom: '0px',
+              left: '0',
+              right: '0',
+              height: '70px',
+              backgroundColor: COLORS.main,
+              display: 'flex',
+              alignItems: 'center',
+              justifyContent: 'space-around',
+              paddingBottom: '12px',
+              paddingTop: '4px',
+              borderTop: `1px solid ${COLORS.main}`,
+              borderRadius: '0 0 38px 38px',
+              zIndex: 0
+            }}>
+              {navigationTabs.map((tab) => (
+                <div
+                  key={tab.id}
+                  onClick={() => navigate(tab.path)}
+                  style={{
+                    display: 'flex',
+                    flexDirection: 'column',
+                    alignItems: 'center',
+                    cursor: 'pointer',
+                    padding: '8px 16px',
+                    borderRadius: '12px',
+                    transition: 'all 0.2s ease',
+                    backgroundColor: isActive(tab.path) ? COLORS.main : 'transparent'
+                  }}
+                >
+                  <img 
+                    src={tab.icon}
+                    alt={tab.label}
+                    style={{
+                      width: '30px',
+                      height: '30px',
+                      marginBottom: '4px',
+                      filter: isActive(tab.path) ? 'none' : 'grayscale(0.5) opacity(0.7)',
+                      transition: 'filter 0.2s ease'
+                    }}
+                  />
+                  <span style={{
+                    fontSize: '15px',
+                    fontWeight: isActive(tab.path) ? '600' : '400',
+                    color: isActive(tab.path) ? COLORS.black : COLORS.gray
+                  }}>
+                    {tab.label}
+                  </span>
+                </div>
+              ))}
+            </div>
+          )}
+          
           {/* 홈 인디케이터 */}
           <div
             style={{
